requirements.txt
streamlit
streamlit-lottie
google-generativeai
sqlalchemy
psycopg2-binary
gTTS
pandas
plotly
pydeck
numpy
geocoder
feedparser
qrcode
requests
pytz